+++ src/main/kotlin/com/autonomousapps/internal/utils/utils.kt
@@ -2,6 +2,9 @@
 
 package com.autonomousapps.internal.utils
 
+import okio.BufferedSource
+import okio.buffer
+import okio.source
 import org.gradle.api.artifacts.result.DependencyResult
 import org.gradle.api.artifacts.result.ResolvedDependencyResult
 import org.gradle.api.file.RegularFile
@@ -18,11 +21,40 @@ internal fun RegularFileProperty.getAndDelete(): File {
   return file
 }
 
-internal inline fun <reified T> RegularFileProperty.fromJsonSet(): Set<T> = get().fromJsonSet()
-internal inline fun <reified T> RegularFile.fromJsonSet(): Set<T> = asFile.readText().fromJsonSet()
+/**
+ * Buffer reads of the nullable RegularFileProperty from disk to the set.
+ */
+internal inline fun <reified T> RegularFileProperty.fromNullableJsonSet(): Set<T> {
+  return orNull?.fromJsonSet() ?: emptySet()
+}
 
+/**
+ * Buffers reads of the RegularFileProperty from disk to the set.
+ */
+internal inline fun <reified T> RegularFileProperty.fromJsonSet(): Set<T> = get().fromJsonSet()
+
+/**
+ * Buffers reads of the RegularFile from disk to the set.
+ */
+internal inline fun <reified T> RegularFile.fromJsonSet(): Set<T> {
+  asFile.bufferRead().use { reader ->
+    return getJsonSetAdapter<T>().fromJson(reader)!!
+  }
+}
+
+/**
+ * Buffers reads of the RegularFileProperty from disk to the set.
+ */
 internal inline fun <reified T> RegularFileProperty.fromJsonList(): List<T> = get().fromJsonList()
-internal inline fun <reified T> RegularFile.fromJsonList(): List<T> = asFile.readText().fromJsonList()
+
+/**
+ * Buffers reads of the RegularFile from disk to the set.
+ */
+internal inline fun <reified T> RegularFile.fromJsonList(): List<T> {
+  asFile.bufferRead().use { reader ->
+    return getJsonListAdapter<T>().fromJson(reader)!!
+  }
+}
 
 internal inline fun <reified K, reified V> RegularFileProperty.fromJsonMapList(): Map<K, List<V>> {
   return get().fromJsonMapList()
@@ -41,23 +73,49 @@ internal inline fun <reified K, reified V> RegularFile.fromJsonMapSet(): Map<K, 
 }
 
 internal inline fun <reified K, reified V> File.fromJsonMapList(): Map<K, List<V>> {
-  return readText().fromJsonMapList()
+  return bufferRead().fromJsonMapList()
 }
 
 internal inline fun <reified K, reified V> File.fromJsonMapSet(): Map<K, Set<V>> {
-  return readText().fromJsonMapSet()
+  return bufferRead().fromJsonMapSet()
 }
 
+/**
+ * Buffers reads of the RegularFileProperty from disk to the set.
+ */
 internal inline fun <reified K, reified V> RegularFileProperty.fromJsonMap(): Map<K, V> = get().fromJsonMap()
+
+/**
+ * Buffers reads of the RegularFile from disk to the set.
+ */
 internal inline fun <reified K, reified V> RegularFile.fromJsonMap(): Map<K, V> = asFile.fromJsonMap()
-internal inline fun <reified K, reified V> File.fromJsonMap(): Map<K, V> = readText().fromJsonMap()
 
+/**
+ * Buffers reads of the File from disk to the set.
+ */
+internal inline fun <reified K, reified V> File.fromJsonMap(): Map<K, V> {
+  bufferRead().use { reader ->
+    return getJsonMapAdapter<K, V>().fromJson(reader)!!
+  }
+}
+
+/**
+ * Buffer reads of the RegularFileProperty from disk to the set.
+ */
 internal inline fun <reified T> RegularFileProperty.fromJson(): T = get().fromJson()
-internal inline fun <reified T> RegularFile.fromJson(): T = asFile.fromJson()
-internal inline fun <reified T> File.fromJson(): T = readText().fromJson()
 
-internal inline fun <reified T> RegularFileProperty.fromNullableJsonSet(): Set<T>? {
-  return orNull?.asFile?.readText()?.fromJsonSet()
+/**
+ * Buffer reads of the RegularFile from disk to the set.
+ */
+internal inline fun <reified T> RegularFile.fromJson(): T = asFile.fromJson()
+
+/**
+ * Buffer reads of the File from disk to the set.
+ */
+internal inline fun <reified T> File.fromJson(): T {
+  bufferRead().use { reader ->
+    return getJsonAdapter<T>().fromJson(reader)!!
+  }
 }
 
 internal fun RegularFileProperty.readLines(): List<String> = get().readLines()
@@ -65,6 +123,8 @@ internal fun RegularFileProperty.readLines(): List<String> = get().readLines()
 internal fun RegularFile.readLines(): List<String> = asFile.readLines()
 
 internal fun RegularFileProperty.readText(): String = get().asFile.readText()
+
+private fun File.bufferRead(): BufferedSource = source().buffer()
 
 // copied from StringsJVM.kt
 internal fun String.capitalizeSafely(locale: Locale = Locale.ROOT): String {