+++ src/main/kotlin/com/autonomousapps/transform/StandardTransform.kt
@@ -27,16 +27,21 @@ internal class StandardTransform(
 
     val declarations = declarations.forCoordinates(coordinates)
 
-    var (mainUsages, testUsages, androidTestUsages) = usages.mutPartitionOf(
+    var (mainUsages, testUsages, androidTestUsages, customJvmUsage) = usages.mutPartitionOf(
       { it.variant.kind == SourceSetKind.MAIN },
       { it.variant.kind == SourceSetKind.TEST },
-      { it.variant.kind == SourceSetKind.ANDROID_TEST }
+      { it.variant.kind == SourceSetKind.ANDROID_TEST },
+      { it.variant.kind == SourceSetKind.CUSTOM_JVM }
     )
-    val (mainDeclarations, testDeclarations, androidTestDeclarations) = declarations.mutPartitionOf(
-      { it.variant(supportedSourceSets)?.kind == SourceSetKind.MAIN },
-      { it.variant(supportedSourceSets)?.kind == SourceSetKind.TEST },
-      { it.variant(supportedSourceSets)?.kind == SourceSetKind.ANDROID_TEST }
-    )
+
+    val hasCustomSourceSets = hasCustomSourceSets(usages)
+    val (mainDeclarations, testDeclarations, androidTestDeclarations, customJvmDeclarations) =
+      declarations.mutPartitionOf(
+        { it.variant(supportedSourceSets, hasCustomSourceSets)?.kind == SourceSetKind.MAIN },
+        { it.variant(supportedSourceSets, hasCustomSourceSets)?.kind == SourceSetKind.TEST },
+        { it.variant(supportedSourceSets, hasCustomSourceSets)?.kind == SourceSetKind.ANDROID_TEST },
+        { it.variant(supportedSourceSets, hasCustomSourceSets)?.kind == SourceSetKind.CUSTOM_JVM }
+      )
 
     /*
      * Main usages.
@@ -63,6 +68,14 @@ internal class StandardTransform(
 
     androidTestUsages = if (isMainVisibleDownstream) mutableSetOf() else reduceUsages(androidTestUsages)
     computeAdvice(advice, androidTestUsages, androidTestDeclarations, androidTestUsages.size == 1)
+
+
+    /*
+     * Custom JVM source sets like 'testFixtures', 'integrationTest' or other custom source sets and feature variants
+     */
+
+    customJvmUsage = reduceUsages(customJvmUsage)
+    computeAdvice(advice, customJvmUsage, customJvmDeclarations, customJvmUsage.size == 1)
 
     return simplify(advice)
   }
@@ -111,9 +124,12 @@ internal class StandardTransform(
     singleVariant: Boolean
   ) {
     val usageIter = usages.iterator()
+    val hasCustomSourceSets = hasCustomSourceSets(usages)
     while (usageIter.hasNext()) {
       val usage = usageIter.next()
-      val declarationsForVariant = declarations.filterToSet { it.variant(supportedSourceSets) == usage.variant }
+      val declarationsForVariant = declarations.filterToSet { declaration ->
+        declaration.variant(supportedSourceSets, hasCustomSourceSets) == usage.variant
+      }
 
       // We have a declaration on the same variant as the usage. Remove or change it, if necessary.
       if (declarationsForVariant.isNotEmpty()) {
@@ -159,7 +175,7 @@ internal class StandardTransform(
             usageIter.remove()
 
             // Don't change a single-usage match, it's correct!
-            if ((!(singleVariant && usage.bucket.matches(theDecl)))) {
+            if (!(singleVariant && usage.bucket.matches(theDecl))) {
               advice += Advice.ofChange(
                 coordinates = coordinates,
                 fromConfiguration = theDecl.configurationName,
@@ -167,6 +183,28 @@ internal class StandardTransform(
               )
             }
           }
+      }
+    }
+
+    // In the very common case that we have one single declaration and one single usage, we have special handling as a
+    // matter of laziness. If the single declaration is both wrong _and_ on a variant, then we transform it to the
+    // correct usage on that same variant. E.g., debugImplementation => debugRuntimeOnly. Without this block, the
+    // algorithm would instead advise: debugImplementation => runtimeOnly.
+    // See `should be debugRuntimeOnly` in StandardTransformTest.
+    if (usages.size == 1 && declarations.size == 1) {
+      val lastUsage = usages.first()
+      if (lastUsage.bucket != Bucket.NONE) {
+        val lastDeclaration = declarations.first()
+        advice += Advice.ofChange(
+          coordinates = coordinates,
+          fromConfiguration = lastDeclaration.configurationName,
+          toConfiguration = lastUsage.toConfiguration(
+            forceVariant = lastDeclaration.variant(supportedSourceSets, hasCustomSourceSets)
+          )
+        )
+
+        // !!!early return!!!
+        return
       }
     }
 
@@ -184,10 +222,13 @@ internal class StandardTransform(
     declarations.asSequence()
       // Don't remove runtimeOnly or compileOnly declarations
       .filterNot { it.bucket == Bucket.COMPILE_ONLY || it.bucket == Bucket.RUNTIME_ONLY }
-      .mapTo(advice) { decl ->
-        Advice.ofRemove(coordinates, decl)
+      .mapTo(advice) { declaration ->
+        Advice.ofRemove(coordinates, declaration)
       }
   }
+
+  private fun hasCustomSourceSets(usages: Set<Usage>) =
+    usages.any { it.variant.kind == SourceSetKind.CUSTOM_JVM }
 
   /** Simply advice by transforming matching pairs of add-advice and remove-advice into a single change-advice. */
   private fun simplify(advice: MutableSet<Advice>): Set<Advice> {
@@ -217,14 +258,31 @@ internal class StandardTransform(
   }
 
   /** e.g., "debug" + "implementation" -> "debugImplementation" */
-  private fun Usage.toConfiguration(): String {
+  private fun Usage.toConfiguration(forceVariant: Variant? = null): String {
     check(bucket != Bucket.NONE) { "You cannot 'declare' an unused dependency" }
 
     fun processor() = if (isKaptApplied) "kapt" else "annotationProcessor"
 
+    fun Variant.configurationNamePrefix(): String = when (kind) {
+      SourceSetKind.MAIN -> variant
+      SourceSetKind.TEST -> "test"
+      SourceSetKind.ANDROID_TEST -> "androidTest"
+      SourceSetKind.CUSTOM_JVM -> variant
+    }
+
+    @OptIn(ExperimentalStdlibApi::class)
+    fun Variant.configurationNameSuffix(): String = when (kind) {
+      SourceSetKind.MAIN -> variant.replaceFirstChar(Char::uppercase)
+      SourceSetKind.TEST -> "Test"
+      SourceSetKind.ANDROID_TEST -> "AndroidTest"
+      SourceSetKind.CUSTOM_JVM -> variant.replaceFirstChar(Char::uppercase)
+    }
+
+    val theVariant = forceVariant ?: variant
+
     if (bucket == Bucket.ANNOTATION_PROCESSOR) {
       val original = processor()
-      return if (variant.variant == Variant.MAIN_NAME) {
+      return if (theVariant.variant == Variant.MAIN_NAME) {
         // "main" + "annotationProcessor" -> "annotationProcessor"
         // "main" + "kapt" -> "kapt"
         if ("annotationProcessor" in original) {
@@ -238,32 +296,34 @@ internal class StandardTransform(
         // "debug" + "annotationProcessor" -> "debugAnnotationProcessor"
         // "test" + "kapt" -> "kaptTest"
         if ("annotationProcessor" in original) {
-          "${configurationNamePrefix()}AnnotationProcessor"
+          "${theVariant.configurationNamePrefix()}AnnotationProcessor"
         } else if ("kapt" in original) {
-          "kapt${configurationNameSuffix()}"
+          "kapt${theVariant.configurationNameSuffix()}"
         } else {
           throw IllegalArgumentException("Unknown annotation processor: $original")
         }
       }
     }
 
-    return if (variant.variant == Variant.MAIN_NAME) {
+    return if (theVariant.variant == Variant.MAIN_NAME && theVariant.kind == SourceSetKind.MAIN) {
       // "main" + "api" -> "api"
       bucket.value
     } else {
       // "debug" + "implementation" -> "debugImplementation"
       // "test" + "implementation" -> "testImplementation"
-      "${configurationNamePrefix()}${bucket.value.capitalizeSafely()}"
+      "${theVariant.configurationNamePrefix()}${bucket.value.capitalizeSafely()}"
     }
   }
 }
 
 private fun Set<Declaration>.forCoordinates(coordinates: Coordinates): Set<Declaration> {
   return asSequence()
-    .filter { it.identifier == coordinates.identifier ||
-      // In the special case of IncludedBuildCoordinates, the declaration might be a 'project(...)' dependency
-      // if subprojects inside an included build depend on each other.
-      (coordinates is IncludedBuildCoordinates) && it.identifier == coordinates.resolvedProject.identifier }
+    .filter {
+      it.identifier == coordinates.identifier ||
+        // In the special case of IncludedBuildCoordinates, the declaration might be a 'project(...)' dependency
+        // if subprojects inside an included build depend on each other.
+        (coordinates is IncludedBuildCoordinates) && it.identifier == coordinates.resolvedProject.identifier
+    }
     // For now, we ignore any special dependencies like test fixtures or platforms
     .filter { !it.doesNotPointAtMainVariant }
     .toSet()
@@ -272,19 +332,6 @@ private fun Set<Declaration>.forCoordinates(coordinates: Coordinates): Set<Decla
 private fun isSingleBucket(usages: Set<Usage>): Boolean {
   return if (usages.size == 1) true
   else usages.mapToSet { it.bucket }.size == 1
-}
-
-private fun Usage.configurationNamePrefix(): String = when (variant.kind) {
-  SourceSetKind.MAIN -> variant.variant
-  SourceSetKind.TEST -> "test"
-  SourceSetKind.ANDROID_TEST -> "androidTest"
-}
-
-@OptIn(ExperimentalStdlibApi::class)
-private fun Usage.configurationNameSuffix(): String = when (variant.kind) {
-  SourceSetKind.MAIN -> variant.variant.replaceFirstChar(Char::uppercase)
-  SourceSetKind.TEST -> "Test"
-  SourceSetKind.ANDROID_TEST -> "AndroidTest"
 }
 
 private fun Sequence<Usage>.filterUsed() = filterNot { it.bucket == Bucket.NONE }