+++ testkit/src/main/kotlin/com/autonomousapps/kit/Dependency.kt
@@ -69,6 +69,11 @@ class Dependency @JvmOverloads constructor(
     }
 
     @JvmStatic
+    fun kotlinTestJunit(configuration: String): Dependency {
+      return Dependency(configuration, "org.jetbrains.kotlin:kotlin-test-junit:$KOTLIN_VERSION")
+    }
+
+    @JvmStatic
     fun scalaStdlib(configuration: String): Dependency {
       return Dependency(configuration, "org.scala-lang:scala-library:2.13.1")
     }
@@ -115,17 +120,17 @@ class Dependency @JvmOverloads constructor(
 
     @JvmStatic
     fun moshi(configuration: String): Dependency {
-      return Dependency(configuration, "com.squareup.moshi:moshi:1.11.0")
+      return Dependency(configuration, "com.squareup.moshi:moshi:1.14.0")
     }
 
     @JvmStatic
     fun moshiKotlin(configuration: String): Dependency {
-      return Dependency(configuration, "com.squareup.moshi:moshi-kotlin:1.11.0")
+      return Dependency(configuration, "com.squareup.moshi:moshi-kotlin:1.14.0")
     }
 
     @JvmStatic
     fun moshiAdapters(configuration: String): Dependency {
-      return Dependency(configuration, "com.squareup.moshi:moshi-adapters:1.11.0")
+      return Dependency(configuration, "com.squareup.moshi:moshi-adapters:1.14.0")
     }
 
     @JvmStatic
@@ -185,17 +190,17 @@ class Dependency @JvmOverloads constructor(
 
     @JvmStatic
     fun dagger(configuration: String): Dependency {
-      return Dependency(configuration, "com.google.dagger:dagger:2.38.1")
+      return Dependency(configuration, "com.google.dagger:dagger:2.44.2")
     }
 
     @JvmStatic
     fun daggerCompiler(configuration: String): Dependency {
-      return Dependency(configuration, "com.google.dagger:dagger-compiler:2.38.1")
+      return Dependency(configuration, "com.google.dagger:dagger-compiler:2.44.2")
     }
 
     @JvmStatic
     fun daggerAndroidCompiler(configuration: String): Dependency {
-      return Dependency(configuration, "com.google.dagger:dagger-android-processor:2.38.1")
+      return Dependency(configuration, "com.google.dagger:dagger-android-processor:2.44.2")
     }
 
     @JvmStatic